+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# 0.0.1
+# Changelog
+
+## 0.0.1
 
 - Implemented basic interface that resembles benchmark.js
 - Implemented statistical calculations.
+
+## 0.0.2
+
+- added `Suite.toJSON` method
+- add `EventType.ALL_COMPLETE` to notify on completion of the `Suite`.